backend/requirements.txt
# Django
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.2
django-redis==5.4.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Production server
gunicorn==22.0.0
uvicorn==0.30.3

# PDF generation
weasyprint==62.3

# Image processing
Pillow==10.4.0

# Utilities
python-decouple==3.8
whitenoise==6.7.0

# Caching
redis==5.0.7
hiredis==3.0.0

# Rate limiting
django-ratelimit==4.1.0
